requirements.txt
pip == 24.0
flask == 2.3.3
pyjwt == 2.8.0
requests-oauthlib == 1.3.1
oauthlib == 3.2.2
stellrent-response == 0.0.4
cryptography == 42.0.5
Flask-HTTPAuth  == 4.8.0
python-decouple == 3.8